Refactor project actions events
