+++ web/modules/projects/projects/src/Event/ProjectApplyEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\projects\Event;
+
+/**
+ * Defines a project apply event.
+ */
+class ProjectApplyEvent extends ProjectEventBase {
+
+}
