+++ web/modules/projects/projects/src/Plugin/rest/resource/ProjectNotifyResource.php
@@ -5,10 +5,6 @@ namespace Drupal\projects\Plugin\rest\resource;
 use Drupal\projects\Entity\Project;
 use Drupal\projects\Event\ProjectNotifyEvent;
 use Drupal\rest\ModifiedResourceResponse;
-use Drupal\rest\Plugin\ResourceBase;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -22,63 +18,7 @@ use Symfony\Component\HttpFoundation\Request;
  *   }
  * )
  */
-class ProjectNotifyResource extends ResourceBase {
-
-  use ProjectActionRestResourceRoutesTrait;
-
-  /**
-   * The event dispatcher.
-   *
-   * @var \Symfony\Component\EventDispatcher\EventDispatcherInterface
-   */
-  protected EventDispatcherInterface $eventDispatcher;
-
-  /**
-   * Constructs a Drupal\rest\Plugin\ResourceBase object.
-   *
-   * @param array $configuration
-   *   A configuration array containing information about the plugin instance.
-   * @param string $plugin_id
-   *   The plugin_id for the plugin instance.
-   * @param mixed $plugin_definition
-   *   The plugin implementation definition.
-   * @param array $serializer_formats
-   *   The available serialization formats.
-   * @param \Psr\Log\LoggerInterface $logger
-   *   A logger instance.
-   * @param \Symfony\Component\EventDispatcher\EventDispatcherInterface $event_dispatcher
-   *   The event dispatcher.
-   */
-  public function __construct(
-    array $configuration,
-    $plugin_id,
-    $plugin_definition,
-    array $serializer_formats,
-    LoggerInterface $logger,
-    EventDispatcherInterface $event_dispatcher
-  ) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition, $serializer_formats, $logger);
-    $this->eventDispatcher = $event_dispatcher;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(
-    ContainerInterface $container,
-    array $configuration,
-    $plugin_id,
-    $plugin_definition
-  ) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->getParameter('serializer.formats'),
-      $container->get('logger.factory')->get('rest'),
-      $container->get('event_dispatcher')
-    );
-  }
+class ProjectNotifyResource extends ProjectActionResourceBase {
 
   /**
    * Responds to POST requests.
@@ -92,7 +32,7 @@ class ProjectNotifyResource extends ResourceBase {
    *   The response.
    */
   public function post(Project $project, Request $request) {
-    $event = new ProjectNotifyEvent($project, $request);
+    $event = new ProjectNotifyEvent($this->currentUser, $project, $request);
     $this->eventDispatcher->dispatch($event);
     return new ModifiedResourceResponse();
   }
